Refatora projeto em arquitetura modular
